+++ Frontend-React/src/components/AddTodoItemForm.tsx
@@ -1,0 +1,53 @@
+import React, { useState, useEffect, ChangeEventHandler } from 'react'
+import { Image, Alert, Button, Container, Row, Col, Form, Table, Stack } from 'react-bootstrap'
+
+const AddTodoItemForm = () => {
+    const [description, setDescription] = useState('');
+    
+    async function handleAdd() {
+        try {
+          alert('todo')
+        } catch (error) {
+          console.error(error)
+        }
+    }
+
+    function handleClear() {
+        setDescription('')
+    }
+
+    const handleDescriptionChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    // todo
+    }
+
+    return (
+        <Container>
+          <h1>Add Item</h1>
+          <Form.Group as={Row} className="mb-3" controlId="formAddTodoItem">
+            <Form.Label column sm="2">
+              Description
+            </Form.Label>
+            <Col md="6">
+              <Form.Control
+                type="text"
+                placeholder="Enter description..."
+                value={description}
+                onChange={handleDescriptionChange}
+              />
+            </Col>
+          </Form.Group>
+          <Form.Group as={Row} className="mb-3 offset-md-2" controlId="formAddTodoItem">
+            <Stack direction="horizontal" gap={2}>
+              <Button variant="primary" onClick={() => handleAdd()}>
+                Add Item
+              </Button>
+              <Button variant="secondary" onClick={() => handleClear()}>
+                Clear
+              </Button>
+            </Stack>
+          </Form.Group>
+        </Container>
+      )
+}
+
+export default AddTodoItemForm
